Generate table from object abstracts with PMID, Gene_ID, and Gene name.
Gets the pmids from the results of the query (abstracts@PMID)
and sends query to pubtator. Formats the results of this query
into a table and eliminates duplicates of pmid+geneID.

This is for working with Gene_ID instead of Symbol, so that the
hypergeometric test can be used.

I still need to link this info with gene symbol for each GeneID.